Add test for calling the main CLI function with correct params
This commit adds a test to verify if the CLI
parameters are being passed correctly to the
Translator object, using a mock. It also fixes
the name of the source API parameter in the main
function, changing it to provider API to match
the parameter name used in the Translator constructor.